Time: 4 ms (89.93%), Space: 21.4 MB (99.06%) - LeetHub

+++ 0907-koko-eating-bananas/0907-koko-eating-bananas.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-int check(vector<int>&piles,int mid)
+long long check(vector<int>&piles,int mid)
 {
     long long  ans=0;
     for(int i=0;i<piles.size();i++)
